+++ packages/material_ui/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
+* Adds an example application demonstrating the package, shown on the pub.dev "Example" tab.
 
 ## 0.0.1